==> notebooks/00-clear data from IDB.ipynb <==
import geopandas as gpd
import pandas as pd
import glob, os
from shapely.geometry import Point
import sidetable

set directiry

g = os.getcwd()
g
path = 'D:\\repos\\CIAT\\IDB_PROJECT'
os.chdir(path)

read excel and check rows and cols

data_raw = pd.read_csv("Data_IDB/Panel_data_PATCA-DR-only_baseline.csv",encoding='LATIN-1' )
data_raw

Points = gpd.read_file("Data_IDB/new/Zona_influencia_DR.shp")
# print(Points.crs)
Points = Points.to_crs("EPSG:4326")
Points.columns

table = Points.stb.freq(['patca'])
table

# names = ['cuest_id_b', 'parcela', 'cultivo','temporary_crop', 'superf_parcela_has', 'Y', 'X', 'patca']
# names = ['patca']

# # making new data frame with dropped NA values 
# new_data = data_raw[names]
# new_data

cords_repY = new_data["Y"].value_counts()
cords_repY
# cords_repX = new_data["X"].value_counts()

Y = cords_repY[cords_repY>1]
print(sum(Y))
print(Y)
# X = cords_repX[cords_repX>1]
Y = Y.index
299-105

cords_repY
633-194-4

df = new_data
for i in Y:
    pos = new_data[new_data["Y"] == i].index
    pos = pos.delete(0)
    df = df.drop(pos)
df

df["cultivo"][df["cultivo"].isnull()]="rest field"
df = df.dropna()
gdf = gpd.GeoDataFrame(
    df, geometry=gpd.points_from_xy(df.X, df.Y))
gdf.to_file("Data_IDB/new/point_raw.shp")
# df[df.Y.isna()]
table = df.stb.freq(['patca'])
table


crs = {'init': 'epsg:4326'}
gdf = gpd.GeoDataFrame(
    df, crs=crs , geometry=gpd.points_from_xy(df.X, df.Y))
gdf.to_file(driver = 'ESRI Shapefile', filename= "Data_IDB/new/new_bl1.shp")

# geometry = [Point(xy) for xy in zip(df["X"], df["Y"])]
# crs = {'init': 'epsg:4326'}
# gdf = gpd.GeoDataFrame(df, crs=crs, geometry = geometry)
# gdf.to_file(driver = 'ESRI Shapefile', filename= "Data_IDB/new/new_bl2.shp")

# df = new_data
# df['geometry'] = df.apply(lambda row: Point(row.X,row.Y),axis=1)
# gdf = gpd.GeoDatFrame(df, geometry = df.geometry)
# gdf.to_file(driver = 'ESRI Shapefile', filename= "Data_IDB/new/new_bl3.shp")

table = df.stb.freq(['cultivo'])
table
table.to_csv("table.csv")

df

pro = gpd.read

435-58
