==> notebooks/03-create database.ipynb <==
import pandas as pd
from pathlib import Path
# import numpy as np

import os,glob
g = os.getcwd()
g
path = 'D:\\repos\\CIAT\\IDB_PROJECT'
os.chdir(path)

path ="Data_IDB/FT"
ext = "csv"
extention = r"{}".format(ext)
type(extention)
direction = Path(r"{}".format(path))
type(direction)
direction
files_list = list(direction.glob(r"*.{}*".format(extention)))
files_list
list_dirs = []
for csv in files_list:
    list_dirs.append(str(csv))
list_dirs

alldb = pd.DataFrame()
for csv in list_dirs:
    file = pd.read_csv(csv)
    alldb = pd.concat([alldb,file],ignore_index = True)

alldb = alldb[["NDVI","OSAVI","EVI","NWVI","ID","patca","date"]]
len(alldb)
alldb

# mean data by ID and date

mean_db = alldb.groupby(["ID","date"],as_index=False ).mean()
mean_db
mean_db.to_csv("Data_IDB/alldf.csv",index=False)
