==> notebooks/09-ANOVA.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
# Display figures inline in Jupyter notebook
import seaborn as sns
import numpy as np
from pathlib import Path
from scipy import signal
import datetime
import numpy as np

import os,glob
g = os.getcwd()
g
path = 'D:\\repos\\CIAT\\IDB_PROJECT'
os.chdir(path)

smooth_all = pd.read_csv("Data_IDB/alldf_smooth.csv")
patca = pd.read_csv("Data_IDB/new/alldf_polygons.csv")
type(smooth_all.ID[0])

ID = str(patca["cuest_id_b"]) + str(patca["parcela"])
str1 = [str(int) for int in patca["cuest_id_b"]] 
str2 = [str(int) for int in patca["parcela"]]
patca["ID"] = [s1 + s2 for s1, s2 in zip(str1, str2)]
patca["ID"] = patca["ID"].astype('int64')
patca = patca[["ID","patca"]]
type(patca.ID[0])
patca

merge = smooth_all.merge(patca, how="inner", on="ID")
merge

mean_M = pd.DataFrame() 
group_id= merge.groupby(["ID"])
for fid in merge.ID.unique():
#     print(fid)
    ID = group_id.get_group(fid)
    ID
    monthly_mean = ID.resample('M').mean()
    # monthly_mean.shift()
    # ############deseason
#     monthly_mean_diff = monthly_mean - monthly_mean.shift()
#     monthly_mean_diff["ID"] = monthly_mean["ID"]
    deseason_trends = pd.concat([deseason_trends,monthly_mean],axis=0)#
    deseason_trends.dropna(inplace=True)

import matplotlib.pyplot as plt
import seaborn as sns
ax = sns.boxplot(x='patca', y='NDVI', data=merge, color='#99c2a2')
ax = sns.swarmplot(x='patca', y='NDVI', data=merge, color='#7d0013')
plt.show()

# load packages
import scipy.stats as stats
# stats f_oneway functions takes the groups as input and returns ANOVA F and p value
fvalue, pvalue = stats.f_oneway(merge['patca'], merge['NDVI'])
print(fvalue, pvalue)
